Fixed incorrect assumption that pid_t is int.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@36071 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ conformance/interfaces/mmap/1-2.c
@@ -30,7 +30,7 @@ int main(void)
 	size_t size = 1024 * 4 * 1024;
 	int fd;
 
-	snprintf(tmpfname, sizeof(tmpfname), "pts_mmap_1_2_%d", getpid());
+	snprintf(tmpfname, sizeof(tmpfname), "pts_mmap_1_2_%ld", (long)getpid());
 
 	/* Create shared object */
 	shm_unlink(tmpfname);
