+++ conformance/interfaces/sem_wait/1-2.c
@@ -31,7 +31,7 @@ int main(void)
 	int value = 10;
 	int val;
 
-	sprintf(semname, "/" FUNCTION "_" TEST "_%d", getpid());
+	sprintf(semname, "/" FUNCTION "_" TEST "_%ld", (long)getpid());
 
 	/* Initial value of Semaphore is 10 */
 	mysemp = sem_open(semname, O_CREAT, 0777, value);
