+++ conformance/interfaces/sem_wait/5-1.c
@@ -28,7 +28,7 @@ int main(void)
 	sem_t *mysemp;
 	char semname[28];
 
-	sprintf(semname, "/" FUNCTION "_" TEST "_%d", getpid());
+	sprintf(semname, "/" FUNCTION "_" TEST "_%ld", (long)getpid());
 
 	/* Initial value of Semaphore is 0 Locked */
 	mysemp = sem_open(semname, O_CREAT, 0777, 0);
